+++ stop-idsvr.sh
@@ -10,11 +10,13 @@ EXAMPLE_NAME='haapi'
 #
 # Download mobile deployment resources
 #
-rm -rf deployment
-git clone https://github.com/curityio/mobile-deployments deployment
-if [ $? -ne 0 ]; then
-  echo 'Problem encountered downloading deployment resources'
-  exit
+if [ ! -d ./deployment ]; then
+  rm -rf deployment
+  git clone https://github.com/curityio/mobile-deployments deployment
+  if [ $? -ne 0 ]; then
+    echo 'Problem encountered downloading deployment resources'
+    exit
+  fi
 fi
 
 #
